+++ lib/Predis/Command/Processor/ProcessorChain.php
@@ -11,6 +11,9 @@
 
 namespace Predis\Command\Processor;
 
+use ArrayAccess;
+use ArrayIterator;
+use InvalidArgumentException;
 use Predis\Command\CommandInterface;
 
 /**
@@ -18,12 +21,12 @@ use Predis\Command\CommandInterface;
  *
  * @author Daniele Alessandri <suppakilla@gmail.com>
  */
-class ProcessorChain implements CommandProcessorChainInterface, \ArrayAccess
+class ProcessorChain implements ArrayAccess, ProcessorInterface
 {
     private $processors = array();
 
     /**
-     * @param array $processors List of instances of CommandProcessorInterface.
+     * @param array $processors List of instances of ProcessorInterface.
      */
     public function __construct($processors = array())
     {
@@ -35,7 +38,7 @@ class ProcessorChain implements CommandProcessorChainInterface, \ArrayAccess
     /**
      * {@inheritdoc}
      */
-    public function add(CommandProcessorInterface $processor)
+    public function add(ProcessorInterface $processor)
     {
         $this->processors[] = $processor;
     }
@@ -43,7 +46,7 @@ class ProcessorChain implements CommandProcessorChainInterface, \ArrayAccess
     /**
      * {@inheritdoc}
      */
-    public function remove(CommandProcessorInterface $processor)
+    public function remove(ProcessorInterface $processor)
     {
         if (false !== $index = array_search($processor, $this->processors, true)) {
             unset($this[$index]);
@@ -71,11 +74,11 @@ class ProcessorChain implements CommandProcessorChainInterface, \ArrayAccess
     /**
      * Returns an iterator over the list of command processor in the chain.
      *
-     * @return \ArrayIterator
+     * @return ArrayIterator
      */
     public function getIterator()
     {
-        return new \ArrayIterator($this->processors);
+        return new ArrayIterator($this->processors);
     }
 
     /**
@@ -109,10 +112,10 @@ class ProcessorChain implements CommandProcessorChainInterface, \ArrayAccess
      */
     public function offsetSet($index, $processor)
     {
-        if (!$processor instanceof CommandProcessorInterface) {
-            throw new \InvalidArgumentException(
+        if (!$processor instanceof ProcessorInterface) {
+            throw new InvalidArgumentException(
                 "A processor chain accepts only instances of ".
-                "'Predis\Command\Processor\CommandProcessorInterface'."
+                "'Predis\Command\Processor\ProcessorInterface'."
             );
         }